Simplified resource classses into functions
