+++ src/core/src/error.rs
@@ -1,5 +1,3 @@
-pub type Result<T> = std::result::Result<T, Error>;
-
 /// The different errors that may happen are stored in this enum. These
 /// include errors specific to some APIs because that way they can be handled
 /// correctly.
@@ -10,6 +8,8 @@ pub enum Error {
     #[error("I/O error: {0}")]
     IO(#[from] std::io::Error),
 }
+
+pub type Result<T> = std::result::Result<T, Error>;
 
 /// For Python interoperability
 impl From<Error> for pyo3::PyErr {
